separate paginate into makeBook and flowIntoBoxes
